Agentic_Design_Patterns/adk/parallelization/pyproject.toml
[project]
name = "parallelization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "google-adk>=1.21.0",
]
